Add Agent UI for chat space
