+++ frontend/src/constants/settings.ts
@@ -1,0 +1,233 @@
+import type { HiveMemoryConfig } from '@/types/config';
+
+// Mock data for development when backend is offline
+export const MOCK_CONFIG: HiveMemoryConfig = {
+  system: {
+    name: 'HiveMemory',
+    version: '0.1.0',
+    debug: true,
+  },
+  logging: {
+    level: 'INFO',
+    format: '%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+    file_path: null,
+    console_output: true,
+  },
+  llm: {
+    gateway: {
+      provider: 'litellm',
+      model: 'deepseek/deepseek-chat',
+      api_key: null,
+      api_base: null,
+      temperature: 0.7,
+      max_tokens: 4096,
+    },
+    librarian: {
+      provider: 'litellm',
+      model: 'deepseek/deepseek-chat',
+      api_key: null,
+      api_base: null,
+      temperature: 0.3,
+      max_tokens: 2048,
+    },
+    worker: {
+      provider: 'litellm',
+      model: 'deepseek/deepseek-chat',
+      api_key: null,
+      api_base: null,
+      temperature: 0.7,
+      max_tokens: 4096,
+    },
+  },
+  embedding: {
+    default: {
+      model_name: 'BAAI/bge-m3',
+      device: 'cpu',
+      cache_dir: null,
+      batch_size: 32,
+      normalize_embeddings: true,
+      dimension: 1024,
+    },
+    perception: {
+      model_name: 'BAAI/bge-m3',
+      device: 'cpu',
+      cache_dir: null,
+      batch_size: 32,
+      normalize_embeddings: true,
+      dimension: 1024,
+    },
+  },
+  qdrant: {
+    host: 'localhost',
+    port: 6333,
+    grpc_port: 6334,
+    api_key: null,
+    collection_name: 'hivememory_main',
+    vector_dimension: 1024,
+    distance_metric: 'Cosine',
+    on_disk_payload: true,
+  },
+  redis: {
+    host: 'localhost',
+    port: 6379,
+    password: null,
+    db: 0,
+    decode_responses: true,
+  },
+  gateway: {
+    interceptor: {
+      enabled: true,
+      enable_system: true,
+      enable_chat: true,
+      custom_system_patterns: [],
+      custom_chat_patterns: [],
+    },
+    analyzer: {
+      enabled: true,
+      context_window: 5,
+      prompt_variant: 'default',
+      prompt_language: 'zh',
+    },
+  },
+  perception: {
+    engine: {
+      type: 'semantic_flow',
+      idle_timeout_seconds: 900,
+      scan_interval_seconds: 30,
+      fold_token_threshold: 32768,
+      fold_retain_recent_blocks: 5,
+      max_resident_topics: 10,
+      relay: {
+        engine: {
+          type: 'simple',
+          max_processing_tokens: 8192,
+        },
+      },
+      adsorber: {
+        semantic_threshold_high: 0.55,
+        semantic_threshold_low: 0.45,
+        short_text_threshold: 0.6,
+        ema_alpha: 0.3,
+        arbiter: {
+          enabled: true,
+          engine: {
+            type: 'reranker',
+            threshold: 0.7,
+          },
+        },
+      },
+    },
+  },
+  generation: {
+    extractor: {
+      enabled: true,
+      system_prompt: null,
+      user_prompt: null,
+    },
+    deduplicator: {
+      enabled: true,
+      high_similarity_threshold: 0.95,
+      low_similarity_threshold: 0.75,
+      content_similarity_threshold: 0.85,
+      enable_vitality_tracking: true,
+    },
+  },
+  retrieval: {
+    renderer: {
+      type: 'full',
+      render_format: 'markdown',
+      max_tokens: 2000,
+      max_content_length: 500,
+      show_artifacts: false,
+      stale_days: 90,
+    },
+    retriever: {
+      type: 'hybrid',
+      top_k: 5,
+      score_threshold: 0.5,
+      enable_parallel: true,
+      dense: {
+        enabled: true,
+        top_k: 10,
+        score_threshold: 0.5,
+        enable_time_decay: true,
+        time_decay_days: 30,
+        enable_confidence_boost: true,
+      },
+      sparse: {
+        enabled: true,
+        top_k: 10,
+        score_threshold: 0.3,
+      },
+      fusion: {
+        type: 'rrf',
+        rrf_k: 60,
+        dense_weight: 1.0,
+        sparse_weight: 0.3,
+        final_top_k: 5,
+      },
+      reranker: {
+        enabled: true,
+        model_name: 'BAAI/bge-reranker-v2-m3',
+        device: 'cpu',
+        use_fp16: false,
+        batch_size: 16,
+        top_k: 5,
+        normalize_scores: true,
+      },
+    },
+  },
+  lifecycle: {
+    high_watermark: 100000,
+    vitality_calculator: {
+      code_snippet_weight: 1.0,
+      fact_weight: 0.9,
+      url_resource_weight: 0.8,
+      reflection_weight: 1.2,
+      user_profile_weight: 1.5,
+      work_in_progress_weight: 1.5,
+      default_weight: 1.0,
+      max_access_boost: 5.0,
+      points_per_access: 0.1,
+      decay_lambda: 0.01,
+    },
+    reinforcement: {
+      enable_event_history: true,
+      event_history_limit: 10,
+      hit_boost: 5.0,
+      citation_boost: 20.0,
+      positive_feedback_boost: 50.0,
+      negative_feedback_penalty: 50.0,
+      negative_confidence_multiplier: 0.8,
+    },
+    archiver: {
+      archive_dir: 'data/archives',
+      compression: true,
+    },
+    garbage_collector: {
+      low_watermark: 20.0,
+      batch_size: 100,
+      enable_schedule: false,
+      interval_hours: 24,
+    },
+  },
+  koakuma: {
+    enabled: true,
+    execution_timeout_seconds: 300,
+    max_recursion_depth: 5,
+    tool_cache_size: 100,
+    python_repl_timeout_seconds: 30,
+    workspace_path: './workspace',
+    file_read_max_bytes: 1048576,
+    file_write_max_bytes: 1048576,
+    web_search_timeout_seconds: 15,
+    mtp_prompt: {
+      enabled: true,
+      language: 'zh',
+      role: 'default',
+      include_demo: true,
+      include_error_handling: true,
+      include_kernel_tools: true,
+    },
+  },
+};
